+++ src/shared/constants/local-storage.constants.ts
@@ -1,1 +1,2 @@
 export const SELECTED_ACCOUNT_KEY = 'selected-account';
+export const EDIT_RECORD_KEY = 'edit-record';