+++ src/shared/types/records.types.ts
@@ -50,6 +50,10 @@ export type ExpenseRecord = AccountRecord & {
   isPaid: boolean;
 }
 
+export type IncomeRecord = AccountRecord & {
+  expensesPaid: ExpensePaid[];
+}
+
 export type ExpensePaid = {
   _id: string;
   shortName: string;
@@ -102,6 +106,24 @@ export type CreateExpensePayload = {
   tag: string[];
   typeOfRecord: 'expense'
 }
+export type EditExpensePayload = CreateExpensePayload & {
+  recordId: string;
+}
+
+export type CreateIncomePayload = {
+  account: string;
+  amount: number;
+  budgets: string[];
+  category: string;
+  date: Date;
+  description: string;
+  expensesPaid: ExpensePaid[]
+  indebtedPeople: IndebtedPeople[];
+  shortName: string;
+  subCategory: string;
+  tag: string[];
+  typeOfRecord: 'income'
+}
 
 export interface CreateExpenseData {
   data: {
@@ -113,7 +135,25 @@ export interface CreateExpenseData {
   version: string;
 }
 
+export interface CreateIncomeData {
+  data: {
+    income: IncomeRecord
+  }
+  error: null;
+  message: string[];
+  success: boolean;
+  version: string;
+}
+
 export interface CreateExpenseError extends Omit<AxiosError, 'response'> {
+  response: AxiosResponse<{
+    error: {
+      message: string;
+    }
+  }>;
+}
+
+export interface CreateIncomeError extends Omit<AxiosError, 'response'> {
   response: AxiosResponse<{
     error: {
       message: string;
@@ -134,6 +174,11 @@ export type CreateExpenseDataForm = {
   description?: string | null | undefined
 }
 
+export type CreateIncomeDataForm = {
+  shortDescription: string
+  description?: string | null | undefined
+}
+
 export type AddIndebtedPeopleDataForm = {
   name: string
 }
@@ -150,7 +195,7 @@ const descriptionValidation = string()
     then: (rule) => rule.min(3, 'Por favor, ingrese una descripción de más de 3 caracteres').max(300, 'Por favor, ingrese una descripción con menos de 300 caracteres.'),
   })
 
-export const CreateExpenseSchema = object().shape({
+export const IncomeExpenseSchema = object().shape({
   shortDescription: shortNameValidation,
   description: descriptionValidation
 }, [